add interactive/non-interactive

+++ pomo_runner.py
@@ -1,4 +1,5 @@
 #! /usr/bin/python3
+import argparse
 from subprocess import call
 from threading import Thread
 
@@ -10,8 +11,6 @@ _kill = False
 
 pomo_state = None
 
-
-####################
 def notify_user(title, message):
     if ENABLE_DESKTOP_NOTIFS:
         try:
@@ -24,9 +23,9 @@ def update_progress_line(td):
     print(timedelta_str(td).ljust(MAX_LINE_LEN), end='\r')
 
 
-def run_stage(stage):
+def run_stage(stage, progress_callback=None):
     global pomo_state
-    t = Thread(target=pomo_state.run, args=(stage, 0.2, update_progress_line))
+    t = Thread(target=pomo_state.run, args=(stage, 0.2, progress_callback))
     t.start()
     try:
         t.join()
@@ -35,22 +34,39 @@ def run_stage(stage):
         t.join()
         raise interrupt
 
+parser = argparse.ArgumentParser(description='A Pomodoro CLI.')
+parser.add_argument('--interactive', dest='interactive', action='store_true')
+args = parser.parse_args()
 
-try:
+if args.interactive:
+    print("Running in interactive mode")
+    try:
+        pomo_state = PomodoroState(ACTIVE_STAGE_MINUTES, REST_STAGE_MINUTES, STATE_FILE)
+        while True:
+            input("Press <enter> to begin a pomodoro.")
+            run_stage(Stage.ACTIVE, progress_callback=update_progress_line)
+            pomo_state.prep_for_rest()
+
+            notify_user("Pomodoro #%d completed" % pomo_state.pomos_completed, "Time for the rest stage")
+            print("Time for the rest stage.")
+            run_stage(Stage.REST, progress_callback=update_progress_line)
+            pomo_state.prep_for_active()
+
+            notify_user("Rest stage #%d completed" % pomo_state.pomos_completed,
+                        "Go to the CLI to start pomodoro #%d" % (pomo_state.pomos_completed + 1))
+            print("Rest stage finished - You've completed %d pomodoro(s)." % pomo_state.pomos_completed)
+    except KeyboardInterrupt:
+        pomo_state.kill()
+    print("Finished - You completed a total of %d pomodoro(s)." % pomo_state.pomos_completed)
+else:
+    # Non-interactive, should be able to just run in the background.
+    print("Running in non-interactive mode")
     pomo_state = PomodoroState(ACTIVE_STAGE_MINUTES, REST_STAGE_MINUTES, STATE_FILE)
     while True:
-        input("Press <enter> to begin a pomodoro.")
         run_stage(Stage.ACTIVE)
         pomo_state.prep_for_rest()
-
-        notify_user("Pomodoro #%d completed" % pomo_state.pomos_completed, "Time for the rest stage")
-        print("Time for the rest stage.")
+        notify_user("Pomodoro completed" % pomo_state.pomos_completed, "Time for the rest stage")
         run_stage(Stage.REST)
         pomo_state.prep_for_active()
+        notify_user("Rest stage completed" % pomo_state.pomos_completed)
 
-        notify_user("Rest stage #%d completed" % pomo_state.pomos_completed,
-                    "Go to the CLI to start pomodoro #%d" % (pomo_state.pomos_completed + 1))
-        print("Rest stage finished - You've completed %d pomodoro(s)." % pomo_state.pomos_completed)
-except KeyboardInterrupt:
-    pomo_state.kill()
-print("Finished - You completed a total of %d pomodoro(s)." % pomo_state.pomos_completed)
